crud stock service

+++ product-service/src/main/java/edu/miu/productservice/repository/ProductRepository.java
@@ -4,10 +4,11 @@ import edu.miu.productservice.entity.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface ProductRepository extends MongoRepository<Product,Long> {
+public interface ProductRepository extends MongoRepository<Product, BigInteger> {
 
     Optional<Product> findByProductNumber(String productNumber);
 }
